+++ x/evidence/types/msgs.go
@@ -3,13 +3,11 @@ package types
 import (
 	"fmt"
 
-	errorsmod "cosmossdk.io/errors"
 	"cosmossdk.io/x/evidence/exported"
 	"github.com/cosmos/gogoproto/proto"
 
 	"github.com/cosmos/cosmos-sdk/codec/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/x/auth/migrations/legacytx"
 )
 
@@ -33,23 +31,6 @@ func NewMsgSubmitEvidence(s sdk.AccAddress, evi exported.Evidence) (*MsgSubmitEv
 	return &MsgSubmitEvidence{Submitter: s.String(), Evidence: any}, nil
 }
 
-// ValidateBasic performs basic (non-state-dependant) validation on a MsgSubmitEvidence.
-func (m MsgSubmitEvidence) ValidateBasic() error {
-	if _, err := sdk.AccAddressFromBech32(m.Submitter); err != nil {
-		return sdkerrors.ErrInvalidAddress.Wrapf("invalid submitter address: %s", err)
-	}
-
-	evi := m.GetEvidence()
-	if evi == nil {
-		return errorsmod.Wrap(ErrInvalidEvidence, "missing evidence")
-	}
-	if err := evi.ValidateBasic(); err != nil {
-		return err
-	}
-
-	return nil
-}
-
 // GetSignBytes returns the raw bytes a signer is expected to sign when submitting
 // a MsgSubmitEvidence message.
 func (m MsgSubmitEvidence) GetSignBytes() []byte {
@@ -63,10 +44,15 @@ func (m MsgSubmitEvidence) GetSigners() []sdk.AccAddress {
 }
 
 func (m MsgSubmitEvidence) GetEvidence() exported.Evidence {
+	if m.Evidence == nil {
+		return nil
+	}
+
 	evi, ok := m.Evidence.GetCachedValue().(exported.Evidence)
 	if !ok {
 		return nil
 	}
+
 	return evi
 }