+++ x/gov/types/v1/msgs_test.go
@@ -21,6 +21,8 @@ var (
 		sdk.AccAddress("test1"),
 		sdk.AccAddress("test2"),
 	}
+
+	metadata = "metadata"
 )
 
 func init() {
@@ -62,7 +64,6 @@ func TestMsgDeposit(t *testing.T) {
 
 // test ValidateBasic for MsgVote
 func TestMsgVote(t *testing.T) {
-	metadata := "metadata" //nolint:goconst
 	tests := []struct {
 		proposalID uint64
 		voterAddr  sdk.AccAddress
@@ -90,7 +91,6 @@ func TestMsgVote(t *testing.T) {
 
 // test ValidateBasic for MsgVoteWeighted
 func TestMsgVoteWeighted(t *testing.T) {
-	metadata := "metadata"
 	tests := []struct {
 		proposalID uint64
 		voterAddr  sdk.AccAddress
@@ -135,7 +135,6 @@ func TestMsgVoteWeighted(t *testing.T) {
 }
 
 func TestMsgSubmitProposal_ValidateBasic(t *testing.T) {
-	metadata := "metadata"
 	// Valid msg
 	msg1, err := v1.NewLegacyContent(v1beta1.NewTextProposal("Title", "description"), addrs[0].String())
 	require.NoError(t, err)