+++ x/nft/msgs.go
@@ -1,10 +1,7 @@
 package nft
 
 import (
-	errorsmod "cosmossdk.io/errors"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 )
 
 const (
@@ -13,28 +10,6 @@ const (
 )
 
 var _ sdk.Msg = &MsgSend{}
-
-// ValidateBasic implements the Msg.ValidateBasic method.
-func (m MsgSend) ValidateBasic() error {
-	if len(m.ClassId) == 0 {
-		return ErrEmptyClassID
-	}
-
-	if len(m.Id) == 0 {
-		return ErrEmptyNFTID
-	}
-
-	_, err := sdk.AccAddressFromBech32(m.Sender)
-	if err != nil {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidAddress, "Invalid sender address (%s)", m.Sender)
-	}
-
-	_, err = sdk.AccAddressFromBech32(m.Receiver)
-	if err != nil {
-		return errorsmod.Wrapf(sdkerrors.ErrInvalidAddress, "Invalid receiver address (%s)", m.Receiver)
-	}
-	return nil
-}
 
 // GetSigners returns the expected signers for MsgSend.
 func (m MsgSend) GetSigners() []sdk.AccAddress {