+++ x/tx/CHANGELOG.md
@@ -33,8 +33,30 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### API Breaking
 
+* [#15709](https://github.com/cosmos/cosmos-sdk/pull/15709):
+  * `GetSignersContext` has been renamed to `signing.Context`
+  * `GetSigners` now returns `[][]byte` instead of `[]string`
+  * `GetSignersOptions` has been renamed to `signing.Options` and requires `address.Codec`s for account and validator addresses
+  * `GetSignersOptions.ProtoFiles` has been renamed to `signing.Options.FileResolver`
+
+## v0.5.1
+
+### Features
+
+* [#15414](https://github.com/cosmos/cosmos-sdk/pull/15414) Add basic transaction decoding support.
+
+## v0.5.0
+
+### API Breaking
+
 * [#15581](https://github.com/cosmos/cosmos-sdk/pull/15581) `GetSignersOptions` and `directaux.SignModeHandlerOptions` now
 require a `signing.ProtoFileResolver` interface instead of `protodesc.Resolver`.
+* [#15742](https://github.com/cosmos/cosmos-sdk/pull/15742) The `direct_aux` package has been renamed to `directaux` in line with Go conventions. No other types were changed during the package rename.
+* [#15748](https://github.com/cosmos/cosmos-sdk/pull/15748) Rename signing.SignerData.ChainId to .ChainID, in line with Go conventions.
+
+### Bug Fixes
+
+* (signing/textual) [#15730](https://github.com/cosmos/cosmos-sdk/pull/15730) make IntValueRenderer.Parse: gracefully handle "" + fuzz
 
 ## v0.4.0
 
@@ -48,3 +70,4 @@ require a `signing.ProtoFileResolver` interface instead of `protodesc.Resolver`.
 
 * [#15302](https://github.com/cosmos/cosmos-sdk/pull/15302) Add support for a custom registry (e.g. gogo's MergedRegistry) to be plugged into SIGN_MODE_TEXTUAL.
 * [#15557](https://github.com/cosmos/cosmos-sdk/pull/15557) Implement unknown field filtering.
+* [#15515](https://github.com/cosmos/cosmos-sdk/pull/15515) Implement SIGN_MODE_LEGACY_AMINO_JSON handler.